fix(core/Canvas): make viewbox/scale cross browser compatible
As it turns out it is virually impossible to get the actual height HTML-
vise from a <svg> element. Seems only possible on Chrome whereas it
returns the BBox (inner size) of the element in other browsers (IE,
Firefox).

To fix this, we are wrapping the element in a <div> which, as part of
the usual dom handling, has correct width / height information available
via div.client(Width|Height).

Fixes the viewbox/scale methods on IE, Firefox and PhantomJS (Webkit).

Related to #29