feat(core): remove commandStack from viewer
This commit removes the command stack from the diagram-js
viewer by making the addShape/addConnection operations
irreversible.

Modeling components must use the diagram-js/lib/cmd module
and implement add/remove/modify commands on top of the exposed
commandStack and the raw APIs provided by the canvas.

This ensures we keep the code base as small and clean as possible.

Related to #41

BREAKING CHANGE:

* Canvas#addShape and Canvas#addConnection are no more undoable via
the commandStack
* The Add(Shape|Connection)Handler has been removed