feat(cmd/CommandStack): add (pre|post)Execute handlers
This commit adds the ability to define preExecute and postExecute
handlers to a command handler. These callbacks are invoked once the
actual command has been executed and may be used to further decorate an
element (or preprocess) it.

The handlers get the context passed.

Closes #44

BREAKING CHANGE:

* The commandStack allows only one command per type to be registered.
This is for simplicity reasons.