feat(overlays): manage position during zoom/scroll/move
This commit enhances the overlays to automatically manage the overlay
position during canvas zoom/scroll as well as shape move.

To do this, the user must now specify a position attribute in the
overlay configuration. It specifies where, relative to the elements top,
left, right or bottom bbox the overlay is attached to the element.

Example:

overlays.add('foo', {
  position: {
    left: 0
  },
  html: '<div>some overlay div</div>'
});

Related to bpmn-io/bpmn-js#54, bpmn-io/bpmn-js#102