feat(core/EventBus): prevent default action via return false
This commit adds a simple way to prevent an events default action (if
specified). Event listeners may simply return false to set the events
isDefaultPrevented() state to false.

This does not stop an events propagation. However, it changes the return
value of eventBus.fire(....) to false, allowing callers to react to it.

Related to #8