feat(command): add CommandStack#canExecute
This adds the #canExecute method to the commandStack as a straight
forward way for components to figure out if a certain interaction with
the diagram is allowed.

Validation during CommandStack#execute does not take place. Instead
users must call CommandStack#canExecute prior to the actual command
execution to figure out whether the action is allowed.

Related to bpmn-io/bpmn-js#137, #8