feat(layout/ManhattanLayout): allow straight lines + keep attachment
This commit updates the manhattan layout component so that it can be
configured to layout straight lines, where possible.

It also does a lot of cleanup on the layout department and ensures we
properly pass change information down to the connection layouters.

* Update original attachment during move
* Allow to layout straight lines (if possible)

* Rename layout/Util -> layout/LayoutUtil
* Rename feature/modeling/Layouter -> layout/BaseLayouter
* Replace BaseLayouter#getConnectionWaypoints -> #layoutConnection

Related to bpmn-io/bpmn-js#249
Related to bpmn-io/bpmn-js#179

BREAKING CHANGES:

This commit renames utilities and updates the Layouter method signature.
See above.