feat(renderer): expose cropping path for shapes + connections
This commit adds the two API methods to the Renderer that expose
the cropping path for shapes and connections:

* Renderer#getShapePath -> defaults to rectangular bbox
* Renderer#getConnectionPath -> defaults to line-segments

Subclasses may overrride the behavior to customize the cropping boxes.

This fixes a downstream bug that pulled the cropping path from
the inserted SVG which is time intensive and error prone (due to lazy
SVG updating we apply).

related to bpmn-io/bpmn-js#236