fix(layout/ManhattanLayout): change straight layouting
When laying out a straight connection we only do it
if the shapes are horizontally or vertically aligned.

Then we compute the orientation to figure out
which axis we are interested in sharing across both
waypoints, assigning the starting point of that axis
to the end point if it's within the target's bounds.

closes #110