feat(Viewer): Open a lightbox on project logo click (optional)
This change is inspired by the discussion at
https://forum.bpmn.io/t/license-questions/85/4

A demo of the effect is available at
http://jsfiddle.net/amatius/vhtn69h8/show/

The viewer (and modeler) can now be initialized with the option
to open a lightbox when the project logo is clicked, rather than
navigating directly to the bpmn.io site.

This behaviour is preferred in some cases to ensure that the user
understands that he/she has left the host application for the
bpmn.io site, thereby reducing the risk that sensitive information
is inadvertantly entered on bpmn.io. It also has the advantage of
making the user explicitly aware of bpmn.io and the relationship to
the host application.
(Yes, the bpmn.io site is opened in a new tab/window, but some
users may find this difficult to understand...)

To opt-in using this feature you must initialize the viewer/modeler
with a new option property

	new BpmnModeler({ poweredBy: 'lightbox' });

As this may be seen as a change to the source code protected by the
license, I've taken extra considerations to fulfill the intention
and purpose (according to my understanding) of the license terms.
For this reason, I had a couple of requirements when designing the
lightbox:

- A clickable link to http://bpmn.io must be included
- The logo of BPMN.IO must be included
- The visual appearance of the BPMN.IO should be recognized in the
  lightbox
- The user must not be discouraged/prevented from navigating to
  http://bpmn.io (but will do so as a consequence of an informed
  decision).

Please note that we are NOT using this change, since this could be
seen as a violation of the license terms until approved by Camunda
(the license holder).