feat(popup-menu): introduce popup menu provider concept
The way how to provide the popup menu with entries has changed.
Now a popup menu provider is needed and has to be registered
initially in the popup menu with an ID. This also changes the
way how a popup menu is created and opened.

BREAKING CHANGE:

 * popupMenu.registerProvider expects a provider which now takes
   care of creating menu entries instead of the old open method
 * popupMenu.create creates the popup menu according to an element
 * popupMenu.open now only expects a position parameter

Related to bpmn-io/bpmn-js#428