Removed fill and stroke from id for markers
The fill and stroke properties are derived from defaultFillColor and defaultStrokeColor. If you use more complex colors for these values (e.g. rgba(x, x, x, x) which is perfectly ok for the color values in the resulting svg), the resulting id contains characters which are not valid in urls (space and comma in this example). And following that, they cannot be used as reference to the markers definition in the svg.