fix(modeling): correct creation of nested lanes
With diagram-js@5 we've introduced the CreateBehavior that ensures
elements are not created on top of lanes but always on top of the actual
participant.

Unfortunately we forgot about the fact that lanes are created once in a
while, too.

This commit accounts for this fact and ensures we do not adjust the
parent of to-be-created lanes.

(A test cases for splitting nested lanes did not exist until now).

Closes #1254
Closes #1253