chore(modeling): consistently create DI with attrs
This fixes the existing DI creation methods in `BpmnFactory` and
simplifies the related `ElementFactory` code that relied on it.

In the past args got ignored and passing attrs to the created DI
was not possible, now it is.

BREAKING CHANGE:

With this change the following `BpmnFactory` API methods got reworked to
take (businessObject, attrs) as an input:

  * `BpmnFactory#createDiEdge`
  * `BpmnFactory#createDiShape`
  * `BpmnFactory#createDiPlane`